--- src/main/java/roomescape/dto/SignupRequest.java
@@ -1,9 +1,0 @@
-package roomescape.dto;
-
-public record SignupRequest(
-        // TODO email 형식 검증
-        String email,
-        String name,
-        String password
-) {
-}
