+++ src/main/java/roomescape/dto/request/CreateReservationThemeRequest.java
@@ -1,10 +1,14 @@
 package roomescape.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
 import roomescape.domain.ReservationTheme;
 
 public record CreateReservationThemeRequest(
+        @NotBlank
         String name,
+        @NotBlank
         String description,
+        @NotBlank
         String thumbnail
 ) {
