+++ src/main/java/roomescape/dto/request/SignupRequest.java
@@ -2,10 +2,14 @@ package roomescape.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
-public record LoginRequest(
+public record SignupRequest(
+        @NotNull
         @Email
         String email,
+        @NotBlank
+        String name,
         @NotBlank
         String password
 ) {
