+++ src/main/java/roomescape/repository/dao/MemberDao.java
@@ -37,6 +37,7 @@ public class MemberDao {
             ps.setString(1, member.name());
             ps.setString(2, member.email());
             ps.setString(3, member.password());
+            ps.setString(4, member.role());
             return ps;
         }, keyHolder);
         Long id = keyHolder.getKey().longValue();
