+++ src/main/java/roomescape/service/MemberService.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import roomescape.auth.JwtProvider;
 import roomescape.domain.Member;
-import roomescape.dto.SignupRequest;
 import roomescape.dto.request.LoginRequest;
+import roomescape.dto.request.SignupRequest;
 import roomescape.dto.response.LoginCheckResponse;
 import roomescape.dto.response.MemberResponse;
 import roomescape.repository.MemberRepository;
